--- Makefile
@@ -1,9 +1,0 @@
-EXES := DuoAdmin.exe
-
-all: $(EXES)
-
-%.exe: %.cs ../Duo.cs
-	dmcs -r:System.Web.Services -r:System.Web.Extensions -r:System.Web $< ../Duo.cs -out:$@
-
-clean:
-	rm -f $(EXES) *~
