--- test/packages.config
@@ -1,20 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="6.0.0" targetFramework="net48" />
-  <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
-  <package id="System.Memory" version="4.5.4" targetFramework="net48" />
-  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net48" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net48" />
-  <package id="System.Text.Encodings.Web" version="6.0.0" targetFramework="net48" />
-  <package id="System.Text.Json" version="6.0.2" targetFramework="net48" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net48" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net48" />
-  <package id="xunit" version="2.4.1" targetFramework="net48" />
-  <package id="xunit.abstractions" version="2.0.3" targetFramework="net48" />
-  <package id="xunit.analyzers" version="0.10.0" targetFramework="net48" />
-  <package id="xunit.assert" version="2.4.1" targetFramework="net48" />
-  <package id="xunit.core" version="2.4.1" targetFramework="net48" />
-  <package id="xunit.extensibility.core" version="2.4.1" targetFramework="net48" />
-  <package id="xunit.extensibility.execution" version="2.4.1" targetFramework="net48" />
-  <package id="xunit.runner.visualstudio" version="2.4.5" targetFramework="net48" developmentDependency="true" />
-</packages>
